+++ package/kedro_viz/data_access/repositories/modular_pipelines.py
@@ -9,8 +9,12 @@ from kedro.pipeline import Pipeline as KedroPipeline
 from kedro.pipeline.node import Node as KedroNode
 
 from kedro_viz.constants import ROOT_MODULAR_PIPELINE_ID
-from kedro_viz.models.flowchart.model_utils import GraphNodeType, ModularPipelineChild
-from kedro_viz.models.flowchart.nodes import GraphNode, ModularPipelineNode
+from kedro_viz.models.flowchart.model_utils import GraphNodeType
+from kedro_viz.models.flowchart.nodes import (
+    GraphNode,
+    ModularPipelineChild,
+    ModularPipelineNode,
+)
 from kedro_viz.utils import _hash, _hash_input_output, is_dataset_param
