+++ package/tests/test_data_access/test_repositories/test_modular_pipelines.py
@@ -6,8 +6,8 @@ from kedro.pipeline.node import Node as KedroNode
 
 from kedro_viz.constants import ROOT_MODULAR_PIPELINE_ID
 from kedro_viz.data_access.repositories import ModularPipelinesRepository
-from kedro_viz.models.flowchart.model_utils import GraphNodeType, ModularPipelineChild
-from kedro_viz.models.flowchart.nodes import ModularPipelineNode
+from kedro_viz.models.flowchart.model_utils import GraphNodeType
+from kedro_viz.models.flowchart.nodes import ModularPipelineChild, ModularPipelineNode
 
 
 @pytest.fixture
